Remove default serverCertNickFile attribute
Previously during installation the server.xml was always created
with a serverCertNickFile attribute. This attribute is supposed
to point to serverCertNick.conf which stores the nickname of the
SSL server cert. If the attribute exists, Tomcat JSS will read
the attribute, load the file, then load the cert. Since the
installation also calls some pki-server commands that depend on
Tomcat JSS, a temporary SSL server cert had to be created early
to prevent failures.

In the future the process might be optimized to reduce/eliminate
the need of a temporary SSL server cert, so the code has been
changed as follows:

- the server.xml will no longer include the attribute by
  default
- the create_temp_sslserver_cert() will be called only if
  the server needs to be started
- the set_sslserver_cert_nickname() will add the attribute
  into the server.xml
- the get_sslserver_cert_nickname() will read the attribute
  from server.xml instead of using a hard-coded value