Remove pki_restart_configured_instance
In the past when installing an additional subsystem (after the
first one) the subsystem had to be started first to execute some
installation steps, then restarted to actually run the subsystem.
The pki_restart_configured_instance param provided an option to
skip the restart to optimize the installation process in IPA.

Now the subsystem no longer needs to be restarted since those
steps can be executed offline, so the code has been modified to
skip the initial startup. The param has been removed as well
since it's no longer used. The DEPRECATED_PARAMS list has also
been updated to generate a removal warning for this param.