Add user and group options for the NSSDatabase class
The user and group options were not used in the NSSDatabase class.
Call chown() for any files/directories created and use runuser
for any external commands if a user is defined.

This is important for filesystem permissions when managing
certificates and private keys inside a software PKCS#11 module
like softHSM.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>