Disable cli return value check when adding certificates
NSS trust is stored in its own token database which needs
to be authenticated to in order to modify trust. certutil has
a bug where tries over and over to authetnicate to a PKCS#11
token to set the trust which inevitably fails because that
isn't where trust needs to be written. So all attempts to
set trust will fail with SEC_ERROR_TOKEN_NOT_LOGGED_IN

Disable the return value check because when adding certs
the trust is ,, anyway so there is nothing to lose.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>