Update CA container to use existing NSS database
pki-ca's Dockerfile has been updated to create a PKI server and
an NSS database just like in pki-server's Dockerfile.

The pki-ca-run has been updated to import the system certs into
the server's NSS database (or generate the certs directly in it),
then create a CA with the certs already in the database just like
in pki-server-run.