Don't use JSS to add the system certificates to the NSS db
When use_uss=True the command pki nss-cert-import is called
which does not provide a password to the generated
certutil command. The end-user must provide it interatively
for the four service certificates during pki-spawn.

When use_jss=False set runas to True so that software-based
tokens (softhsm) will generate correct ownership.

Utilize the tmpdir provided by the class to remove the
try/finally block.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>