Fix Azure pipeline
Due to recent changes in Azure or Fedora it's no longer possible
to run tests with the standard Fedora image.

To fix the problem, the Dockerfile has been updated to define a
new Fedora image that provides a systemd service. The Azure job
has been updated to build the image, create a container with the
image, then run the tests with it. The job has also been updated
to test creating a server, starting it up, checking the status,
shutting it down, and removing the server.