Enable role-based security for REST API
Generally the access to the REST API is controlled by the
legacy ACL defined in acl.ldif which is imported into the
LDAP database during installation. Since there is no
automatic database upgrade mechanism, it's not easy to
update the ACL in existing instances.

To avoid the issue, the web.xml files have been modified
to enable role-based security for REST API. This way the
ACL for new REST API can be defined using the standard
security annotations such as @RolesAllowed. The legacy ACL
will continue to work, but in the future they might be
migrated into role-based security as well.