Move the realm developed for ACME to the server package
ACME subsystem contains a set of realm working with the ProxyRealm and
allowing to configure different authentication models.
These realm are moved outside of ACME package so they can be used by
other subsystems (e.g. EST).
Additionally, these realm have been updated to extend the tomcat
RealmBase so they do not need to be wrapped to be used with the
ProxyRealm.
Finally, the InMemoryRealm supports multiple roles.

However, the configuration model is different from the container managed
realm so, at the moment, they cannot replace the tomcat realm to use at
container level.

To avoid risk and extra QE work, this commit DOES NOT update the
ACME classes.  This results in some code duplication, but this
commit is only intended for the v10.13 release branch.  In the
mainline branches, the ACME realm classes WERE refactored to use the
new realm classes.