Add token to the NSS nickname when modifying certificate trust
The certificate may only be visible to to the token when standing
up a clone. In order for the trust to be set certutil must be
able to read it. Even though trust is stored in the softokn being
able to see that the certificate exists is enough to proceed.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>