Provide all passwords to certutil when modifying certificate trust
The certutil -f option supports a password file in the form of
token:password.

It will seek out the correct password for a token when a login
is required. This works in conjunction with
https://bugzilla.mozilla.org/show_bug.cgi?id=1782980 so
trust can be added to a certificate that is only visible on a
token.

Certificate trust is stored in the softokn and not in a PKCS#11
token so both passwords may be necessary when modifying/adding trust.
One to find the certificate on the token and the other to write to
the NSS database.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>