Split SubsystemRangeUpdateCLI
The SubsystemRangeUpdateCLI has been split into CARangeUpdateCLI
and KRARangeUpdateCLI to avoid unnecessarily updating ID ranges
if the server is configured with RSNv3.

The CARangeUpdateCLI.updateSerialNumberRange() will update the
cert ID range only if the CA is configured with a legacy cert ID
generator.

The KRARangeUpdateCLI.updateSerialNumberRange() will update the
key ID range only if the KRA is configured with a legacy key ID
generator.

The SubsystemRangeUpdateCLI.updateRequestNumberRange() will
update the cert/key request ID range only if the CA/KRA is
configured with a legacy request ID generator.