Add test for installing TPS with HSM
A new test has been added to verify TPS installation with HSM.
In this case TPS system certs and keys will be created in HSM
except for the subsystem and SSL server certs which are already
created during CA installation. The TPS system certs will also
exist in the internal token but without the keys. Certs that
have a key in the same token can be identified by the "u" trust
flags.