Add new --attributes option to <subsystem>-user-add
Originally, this was to set nsPagedSizeLimit=20000 on creation of users

Currently every entry is pulled by default which has been causing users
problems. There is an RFE to change this attribute dynamically based on
the size of the backend DB but for now set a reasonable default.

Instead of just adding this one attribute, a general method for adding
arbitrary single-valued attributes on user creation is introduced.

Resolves: #1959057