Fix cert rest API with json
Rest APIs to enroll and revoke certificates using CLI and json format
were broken in several points.

For the enrollment the problem was with the policy `id`. It was not
included in the lis json sent to the client for review breaking the
client. It is now included.
For the revocation there was a name mismatch between client and server.

Additionally, two test files have been fixed. The one for the enrollment
has a new CSR because the old one made use of obsolete and not supported
algorithm and the other for the revocation has a fixed on the time
counter.

Fix https://bugzilla.redhat.com/show_bug.cgi?id=2053189