Fix adding Profile/Profile Mapping in TPS UI
There were some naming mismatches and request format issues which are
now resolved.

In addition, there was a larger problem that JS arrays are not always
deserialised correctly, so a new StringHashMapValueDeserializer class is
introduced to assist with this.

Adding ProfileMappingData objects still did not work, as the Id and
ProfileMappingID were used interchangeably causing a null reference.
ProfileMappingData has been modified to be of the same format as
ProfileData.