Introduce pki-rpminspect profile for rpminspect
Allows more fine-grained control of checks than CLI arguments, which
allows us to run the javabytecode, runpath and annocheck jobs again
without disabling them completely.

The license job is disabled because of a bug in rpminspect itself.

The metadata job is still disabled, but we can probably now configure it
to pass using this profile.