Rename pki-server sd-host-add to pki-server sd-subsystem-add
The pki-server sd-host-add command has been renamed into
pki-server sd-subsystem-add since it's used to manage subsystems
(which are identified by hostname, port, and subsystem type)
in the security domain instead of hosts.

The original CLI was only used for development and has never
been officially documented so it's not necessary to document
this change.