Drop PKISocketFactory dependency on CMS.getCMSEngine()
The PKISocketFactory.init() has been modified to use the defaults
if there is no configuration provided. The "!external" expression
is supposed to indicate cases where the code is used within PKI
server, but in those cases the configuration is always provided,
so it's not necessary to use the expression anymore. This change
removes the dependency on CMS.getCMSEngine().