Refactor HttpClient to not create unnecessary sockets and streams
The current implementation creates a set of null streams and sockets,
then decides part way through the send process whether to use an SSL
socket or not, creating streams from the appropriate socket accordingly.
This is incompatible with using try-with-resources as you cannot
reassign to a controlled resource. Also, it is wasteful and confusing
that the unnecessary objects exist (but understandable as the code
predates the existence of try-with-resources).

Separate SSL/non-SSL send methods are created, which only create the
necessary streams and sockets for that operation. New helper methods are
introduced to send/handle the request/response.