Provide testing configuration for nss SSL Connection session management.
Currently our pki servers make ssl connections to various directory servers and other
pki subsystems. We use the jss bindings to nss to manage these connections.

Often we are interested in observing the actions taken by servers to verify various client
and server certificates as they move through the system when connections take place.

One of the methods is to observe requests to an OCSP responder if the certs are configured with AIA extensions.

As nss manages connections between various end points, it keeps a cache of ssl sessions for both
client ssl sessions and server ssl sessions.

If a session or ssl connection made is determined to be in the cache, there is no full handshake done.
The effect of this would be that not every connection will result in a verification of server or client certs.

Often testing requires that we might want to temporarilly configure pki servers to mitigate the effect of this
caching and cause more cert verifications to take place.

Therefore this fix will allow the following hidden config parameters for any pki subsystem:

jss.ssl.server.invalidateSessionAfterHandshake=true
jss.ssl.client.invalidateSessionAfterHandshake=true

Note these setttings default to "false" when not present, so they should only be used in rare testing
scenarios, since the settings will result in unusual and expensive cert verifications that nss itself
doesn't deem necessary.

If the first setting is set to true, every time an SSL server socket connection has its handhsake completed,
the code will explcitly invalidate the corresponding communication session from the server session cache in nss.

This means the next time a similar connection is attempted, (such as subystem to subsytem communications ex: ca > kra),
(even with a new socket), nss will not know about the session and will initiate
the full SSL handshake which includes cert verification calls.

The second setting is very similiar but only for client sockets and sessions. There are two settings if one would
want to pick and choose to enable one or the other or both, or neither.