Fix JUnit tests before attempting to migrate to JUnit5
Most of the failing tests were due to a SecureRandom object not
instantiating correctly due to the Mozilla-JSS security provider not
being set up for the test. Only one test actually needs the SecureRandom
so it is removed from the test base class and passed null into the
method that requires it.

One DBRegistryTest fails because it can't find a logging bundle during
the test. The log message generated only happens during the failure
condition the test is checking for, so for now catch that exception (the
test will need refactoring anyway).