Add SAN option for pki nss-cert commands
The pki nss-cert-request and pki nss-cert-issue commands have
been modified to provide an option to specify a SAN extension
when generating a CSR and issuing a certificate which will make
it easier to test SAN extensions.

The NSSExtensionGenerator has been modified to support `critical`
option and specific DNS names for SAN extension.

The CI test has been updated to validate the new option.