Refactor com.netscape.test.TestRunner to use the JUnit 5 runner
The test launcher does automatic test discovery so we can drop the
fragile bash script that does it in cmake, which simplifies things.

This converts 103 of the 166 tests, which were the "easy" ones. The
remaining tests all extend CMSBaseTestCase which is a subclass of
JUnit's TestCase, so they will need to be reimplemented separately.