Fix bug in AgentCertService.revokeCert
The current implementation always gets the CertRecord from the
RevocationProcessor instead of the provided clientRecord as the id
comparison was done between two different types, hence is always false.

The fix compares the id's as BigInteger types, and checks the
clientRecord for null first as it is dereferenced later without a null
check and is instantiated null.