Extract the waiting for build step into a new workflow
This simplifies the top-level test files slightly. Also, I added a check
to pull the images at the end of the wait. This way, the job fails if
the build job is not finished. Currently the wait job always succeeds,
then all subsequent jobs fail as a result. We can use this fail
condition to trigger a retry.