Introduce a retry to wait-for-build.yml
If the waiting-for-build-job fails, run a second job. This should allow
multiple CI suites to run simultaneously again. Additional runs can be
added if needed. This is a bit of a dirty hack, but GitHub actions
doesn't allow loop control flow and there is no way to extend the
timeout of the build job.