Fix bug in TPSConnectorClient.getConnector()
Previously, if connectors.getTotal() < 1 then we threw
ResourceNotFoundException. Then we proceed to try access the next
element in the connectors.getEntries() iterator.

The issue here is that DataCollection offers no guarantee that the total
field is equivalent to connectors.getEntries().size(), it is left to
calling code to update the total. This causes NoSuchElementException to
be thrown on TPS startup. The connector not being present is actually
expected by the calling code, so that is fine, but it is ugly that there
is a stack trace associated with expected behaviour.

We now check connectors.getEntries().isEmpty() instead, throwing the
correct ResourceNotFoundException, which is properly handled by the
calling code.