Migrate EST tests to ansible
Tests based on GitHub workflow cannot easily be executed locally
before the code is pushed making more difficult to identify problems
during the development.

Additionally, there are several limitation on how workflows can be
combined.

Moving to ansible should solve the limitation and allow the
execution of test on developer machine.

This is the first step aims at converting a single workflow to ansible.

The preliminary steps of building the docker images and deploy onto the
runner are still based on github actions.

If/when all workflows will use ansible the overall CI activities could
be reorganised to optimise the execution in the available runners.