Update POM files
The POM files have been modified to use the latest dependencies.
The dependencies have been moved into the modules that actually
need them. The maven-surefire-plugin has been added to run JUnit
5 tests properly. A new POM file has also been added for EST.

The RPM spec and Azure pipelines have been updated to use the
same dependencies as in the POM files.