For unknown certificates OCSP should have unknown CertStatus
For unknown certificates OCSP should have unknown CertStatus (part 2)

The CA's internal OCSP fails to handle certs issued by an unknown CA.
There is code in the CA's validation to handle that scenario but that
validation is never triggered as the request handling code that wraps it
considers not knowing the origin CA to be an error condition.

The code is changed to allow the validating CA to proceed even if the
origin CA is unknown, reporting Unknown for the CertStatus, while
delegating to the origin CA if it is found.

Internal OCSP CA identification using request hash

In case of multiple CAs the correct one was selected using the first
certificate. This could provide inconsistent. Now the selection is based
on the request issuer name.

Additionally, the output has been made consistent with the external OCSP
for all the possibilities of subject and issuers.

Bug2221818-ocsp-unknownCA-addendum

a couple things:
- respond with Unknown when CA can't find cert in its db
- added safety net Exception to return Unknown
- minor debug message when CA signs an object that's not a cert

fixes https://bugzilla.redhat.com/show_bug.cgi?id=2221818

Bug2221818-ocsp-unknownCA-addendum2

minor adjustment for Exception and debug messages.

Fix OCSP verification of requests hashes

OCSP requests have the certificate serial number, the hash of CA DN and
the hash of the CA public key. According to the specification, in order
to recognise a request both hashes have to match but the current implementation
was verifying only the public key hash.

This commit add a check on the other hash of the request.

FIx OCSP DefStore for unknown CA error

Fix OCSP ldap responder output with correct logic

The OCSP responder was failing if CA or CRL were not found. This has been
modified so the responder will provide an `unknown` result in these
cases.

Improve exception handling for OCSP validation