Create temporary files to be shared between uses in /tmp
Some commands need to be executed as the pki user and not
root to retain filesystem permissions. There are a few
places where passwords are written to files as root to be
passed into commands executed by pkiuser.

If a private temporary directory is set before pkispawn
is called then this method for sharing passwords between
users will not work because the file will be unreadable.

So force these calls to use /tmp directly instead of the
private temporary directory.

Fixes: https://github.com/dogtagpki/pki/issues/4475

Signed-off-by: Rob Crittenden <rcritten@redhat.com>