Add Implementation-Version to Manifests via pom.xml
The /getStatus endpoints in the web UI were returning a null
version since the build was switched to Maven.

It is probably now possible to remove the individual manifest files
but I have left them as they are for now so we can still build with
cmake if we wish.